+++ src/Command/TokenRevokeCommand.php
@@ -3,15 +3,19 @@
 namespace App\Command;
 
 use App\Helper\Regex;
+use App\Security\AuthProvider;
 use App\Service\ElementManager;
 use App\Style\EmberNexusStyle;
+use App\Type\TokenRevokeEntry;
 use App\Type\TokenStateType;
 use EmberNexusBundle\Service\EmberNexusConfiguration;
 use Exception;
 use Laudis\Neo4j\Databags\Statement;
 use Laudis\Neo4j\Types\DateTimeZoneId;
+use Predis\Client as RedisClient;
 use Psr\Log\LoggerInterface;
 use Ramsey\Uuid\Uuid;
+use Ramsey\Uuid\UuidInterface;
 use Safe\DateTime;
 use Symfony\Component\Console\Attribute\AsCommand;
 use Symfony\Component\Console\Command\Command;
@@ -20,7 +24,6 @@ use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Console\Question\ConfirmationQuestion;
-use Symfony\Component\Console\Style\OutputStyle;
 use Syndesi\CypherEntityManager\Type\EntityManager as CypherEntityManager;
 
 use function Safe\preg_match;
@@ -31,7 +34,7 @@ use function Safe\preg_match;
 #[AsCommand(name: 'token:revoke', description: 'Revoke tokens.')]
 class TokenRevokeCommand extends Command
 {
-    private OutputStyle $io;
+    private EmberNexusStyle $io;
 
     public const OPTION_FORCE = 'force';
     public const OPTION_DRY_RUN = 'dry-run';
@@ -45,10 +48,22 @@ class TokenRevokeCommand extends Command
     public const NUMBER_OF_TOKENS_TO_DISPLAY = 10;
     public const TOKEN_REVOKE_BATCH_SIZE = 10;
 
+    private bool $isForceActive = false;
+    private bool $isDryRunActive = false;
+    private bool $isListAllAffectedTokensActive = false;
+    private ?UuidInterface $userUuid = null;
+    private ?UuidInterface $groupUuid = null;
+    private ?DateTime $issuedBefore = null;
+    private ?DateTime $issuedAfter = null;
+    private bool $isFilterIssuedWithoutExpirationDateActive = false;
+    private bool $isFilterIssuedWithExpirationDateActive = false;
+
     public function __construct(
         private ElementManager $elementManager,
         private CypherEntityManager $cypherEntityManager,
+        private RedisClient $redisClient,
         private EmberNexusConfiguration $emberNexusConfiguration,
+        private AuthProvider $authProvider,
         private LoggerInterface $logger
     ) {
         parent::__construct();
@@ -127,38 +142,72 @@ class TokenRevokeCommand extends Command
 
         $this->io->title('Token Revocation');
 
-        $filters = [];
-        $arguments = [];
+        $this->validateInput($input);
+        $tokenRevokeEntries = $this->getAffectedTokenRevokeEntries();
 
-        if (
-            true === $input->getOption(self::OPTION_ISSUED_WITH_EXPIRATION_DATE)
-            && true === $input->getOption(self::OPTION_ISSUED_WITHOUT_EXPIRATION_DATE)
-        ) {
+        if (0 === count($tokenRevokeEntries)) {
+            $this->io->finalMessage('No active tokens found.');
+
+            return Command::SUCCESS;
+        }
+
+        $this->printAffectedTokens($tokenRevokeEntries);
+
+        if (true === $input->getOption(self::OPTION_DRY_RUN)) {
+            $this->io->finalMessage('Dry run finished.');
+
+            return Command::SUCCESS;
+        }
+
+        if (true !== $this->isForceActive) {
+            /**
+             * @var QuestionHelper $helper
+             */
+            $helper = $this->getHelper('question');
+            $question = new ConfirmationQuestion(sprintf(
+                '  Are you sure you want to revoke all %d tokens? [y/N]: ',
+                count($tokenRevokeEntries)
+            ), false);
+            if (!$helper->ask($input, $output, $question)) {
+                $this->io->finalMessage('Aborted revoking tokens.');
+
+                return Command::FAILURE;
+            }
+            $this->io->newLine();
+        }
+
+        $this->io->writeln('  Revoking tokens...');
+
+        $this->revokeTokens($tokenRevokeEntries);
+
+        $this->io->finalMessage('Successfully revoked tokens.');
+
+        return Command::SUCCESS;
+    }
+
+    private function validateInput(InputInterface $input): void
+    {
+        $this->isForceActive = $input->getOption(self::OPTION_FORCE);
+        $this->isDryRunActive = $input->getOption(self::OPTION_DRY_RUN);
+        $this->isListAllAffectedTokensActive = $input->getOption(self::OPTION_LIST_ALL_AFFECTED_TOKENS);
+
+        $this->isFilterIssuedWithExpirationDateActive = $input->getOption(self::OPTION_ISSUED_WITH_EXPIRATION_DATE);
+        $this->isFilterIssuedWithoutExpirationDateActive = $input->getOption(self::OPTION_ISSUED_WITHOUT_EXPIRATION_DATE);
+
+        if ($this->isFilterIssuedWithExpirationDateActive && $this->isFilterIssuedWithoutExpirationDateActive) {
             throw new Exception(sprintf('Using both %s and %s is not possible.', self::OPTION_ISSUED_WITH_EXPIRATION_DATE, self::OPTION_ISSUED_WITHOUT_EXPIRATION_DATE));
         }
-        if (true === $input->getOption(self::OPTION_ISSUED_WITH_EXPIRATION_DATE)) {
-            $filters[] = 't.expirationDate IS NOT NULL';
-        }
-        if (true === $input->getOption(self::OPTION_ISSUED_WITHOUT_EXPIRATION_DATE)) {
-            $filters[] = 't.expirationDate IS NULL';
-        }
 
-        $optionIssuedBefore = null;
         if (null !== $input->getOption(self::OPTION_ISSUED_BEFORE)) {
-            $optionIssuedBefore = DateTime::createFromFormat('Y-m-d H:i', $input->getOption(self::OPTION_ISSUED_BEFORE));
-            $filters[] = 't.created < $issuedBefore';
-            $arguments['issuedBefore'] = $optionIssuedBefore;
+            $this->issuedBefore = DateTime::createFromFormat('Y-m-d H:i', $input->getOption(self::OPTION_ISSUED_BEFORE));
         }
 
-        $optionIssuedAfter = null;
         if (null !== $input->getOption(self::OPTION_ISSUED_AFTER)) {
-            $optionIssuedAfter = DateTime::createFromFormat('Y-m-d H:i', $input->getOption(self::OPTION_ISSUED_AFTER));
-            $filters[] = 't.created > $issuedAfter';
-            $arguments['issuedAfter'] = $optionIssuedAfter;
+            $this->issuedAfter = DateTime::createFromFormat('Y-m-d H:i', $input->getOption(self::OPTION_ISSUED_AFTER));
         }
 
-        if ($optionIssuedBefore && $optionIssuedAfter) {
-            if ($optionIssuedBefore->getTimestamp() < $optionIssuedAfter->getTimestamp()) {
+        if ($this->issuedBefore && $this->issuedAfter) {
+            if ($this->issuedBefore->getTimestamp() < $this->issuedAfter->getTimestamp()) {
                 throw new Exception(sprintf('%s can not be before %s.', self::OPTION_ISSUED_BEFORE, self::OPTION_ISSUED_AFTER));
             }
         }
@@ -166,26 +215,68 @@ class TokenRevokeCommand extends Command
         if (null !== $input->getOption(self::OPTION_USER)) {
             $userIdentifier = $input->getOption(self::OPTION_USER);
             if (preg_match(Regex::UUID_V4, $userIdentifier)) {
-                $filters[] = 'u.id = $userIdentifier';
+                $this->userUuid = Uuid::fromString($userIdentifier);
             } else {
-                $filters[] = sprintf(
-                    'u.%s = $userIdentifier',
-                    $this->emberNexusConfiguration->getRegisterUniqueIdentifier()
-                );
+                $res = $this->cypherEntityManager->getClient()->runStatement(new Statement(
+                    sprintf(
+                        'MATCH (u:User {%s: $userIdentifier}) RETURN u.id LIMIT 1',
+                        $this->emberNexusConfiguration->getRegisterUniqueIdentifier()
+                    ),
+                    [
+                        'userIdentifier' => $userIdentifier,
+                    ]
+                ));
+
+                if (1 !== count($res)) {
+                    throw new Exception(sprintf('Unable to find user identified by %s.', $userIdentifier));
+                }
+
+                $this->userUuid = Uuid::fromString($res[0]['u.id']);
             }
-            $arguments['userIdentifier'] = $userIdentifier;
         }
 
         if (null !== $input->getOption(self::OPTION_GROUP)) {
+            $this->groupUuid = Uuid::fromString($input->getOption(self::OPTION_GROUP));
+        }
+    }
+
+    private function buildGetAffectedTokenRevokeEntriesQuery(): Statement
+    {
+        $filters = [];
+        $arguments = [];
+
+        if ($this->isFilterIssuedWithExpirationDateActive) {
+            $filters[] = 't.expirationDate IS NOT NULL';
+        }
+        if ($this->isFilterIssuedWithoutExpirationDateActive) {
+            $filters[] = 't.expirationDate IS NULL';
+        }
+
+        if ($this->issuedBefore) {
+            $filters[] = 't.created < $issuedBefore';
+            $arguments['issuedBefore'] = $this->issuedBefore;
+        }
+
+        if ($this->issuedAfter) {
+            $filters[] = 't.created > $issuedAfter';
+            $arguments['issuedAfter'] = $this->issuedAfter;
+        }
+
+        if ($this->userUuid) {
+            $filters[] = 'u.id = $userIdentifier';
+            $arguments['userIdentifier'] = $this->userUuid->toString();
+        }
+
+        if ($this->groupUuid) {
             $filters[] = '(t)<-[:OWNS]-(:User)-[:IS_IN_GROUP*1..]->(:Group {id: $groupIdentifier})';
-            $arguments['groupIdentifier'] = $input->getOption(self::OPTION_GROUP);
+            $arguments['groupIdentifier'] = $this->groupUuid->toString();
         }
 
         $joinedFilters = join("\nAND ", $filters);
         $finalQuery = sprintf(
             "MATCH (t:Token {state: '%s'})<-[:OWNS]-(u:User)\n".
             '%s%s%s'.
-            "RETURN t.id, t.created, t.expirationDate, u.id, u.%s as userUniqueIdentifier\n".
+            "RETURN t.id, t.created, t.expirationDate, t.hash, u.id, u.%s as userUniqueIdentifier\n".
             'ORDER BY t.created ASC, t.id ASC',
             TokenStateType::ACTIVE->value,
             count($filters) > 0 ? 'WHERE ' : '',
@@ -194,60 +285,78 @@ class TokenRevokeCommand extends Command
             $this->emberNexusConfiguration->getRegisterUniqueIdentifier()
         );
 
-        $res = $this->cypherEntityManager->getClient()->runStatement(
-            new Statement($finalQuery, $arguments)
+        return new Statement($finalQuery, $arguments);
+    }
+
+    /**
+     * @return TokenRevokeEntry[]
+     */
+    private function getAffectedTokenRevokeEntries(): array
+    {
+        $queryResultLines = $this->cypherEntityManager->getClient()->runStatement(
+            $this->buildGetAffectedTokenRevokeEntriesQuery()
         );
-
-        $this->io->isDecorated();
-
-        $countTokensToBeRevoked = count($res);
-        if (0 === $countTokensToBeRevoked) {
-            $this->io->finalMessage('No active tokens found.');
-
-            return Command::SUCCESS;
+        $result = [];
+        foreach ($queryResultLines as $queryResultLine) {
+            $tokenCreated = $queryResultLine['t.created'];
+            $expirationDate = $queryResultLine['t.expirationDate'];
+            /**
+             * @var DateTimeZoneId  $tokenCreated
+             * @var ?DateTimeZoneId $expirationDate
+             */
+            $result[] = new TokenRevokeEntry(
+                Uuid::fromString($queryResultLine['t.id']),
+                $tokenCreated->toDateTime(),
+                $expirationDate?->toDateTime(),
+                Uuid::fromString($queryResultLine['u.id']),
+                $queryResultLine['userUniqueIdentifier'],
+                $queryResultLine['t.hash']
+            );
         }
 
-        $showOnlyFirstTokens = $countTokensToBeRevoked > self::NUMBER_OF_TOKENS_TO_DISPLAY
-            && false === $input->getOption(self::OPTION_LIST_ALL_AFFECTED_TOKENS)
-            && false === $input->getOption(self::OPTION_DRY_RUN);
+        return $result;
+    }
 
-        if (true === $input->getOption(self::OPTION_DRY_RUN)) {
+    /**
+     * @param TokenRevokeEntry[] $tokenRevokeEntries
+     */
+    private function printAffectedTokens(array $tokenRevokeEntries): void
+    {
+        $countTokenRevokeEntries = count($tokenRevokeEntries);
+
+        $showOnlyFirstTokens = $countTokenRevokeEntries > self::NUMBER_OF_TOKENS_TO_DISPLAY
+            && false === $this->isListAllAffectedTokensActive
+            && false === $this->isDryRunActive;
+
+        if (true === $this->isDryRunActive) {
             $this->io->writeln(sprintf(
                 '  %d tokens would be affected by revocation:',
-                $countTokensToBeRevoked
+                $countTokenRevokeEntries
             ));
         } else {
             if ($showOnlyFirstTokens) {
                 $this->io->writeln(sprintf(
                     '  %d tokens are affected by revocation. First %s are shown:',
-                    $countTokensToBeRevoked,
+                    $countTokenRevokeEntries,
                     self::NUMBER_OF_TOKENS_TO_DISPLAY
                 ));
             } else {
                 $this->io->writeln(sprintf(
                     '  %d tokens are affected by revocation:',
-                    $countTokensToBeRevoked
+                    $countTokenRevokeEntries
                 ));
             }
         }
         $this->io->newLine();
 
         $rows = [];
-        foreach ($res as $tokenResult) {
-            $tokenCreated = $tokenResult['t.created'];
-            if ($tokenCreated instanceof DateTimeZoneId) {
-                $tokenCreated = $tokenCreated->toDateTime()->format('Y-m-d H:i:s');
-            }
-            $tokenExpires = $tokenResult['t.expirationDate'];
-            if ($tokenExpires instanceof DateTimeZoneId) {
-                $tokenExpires = $tokenExpires->toDateTime()->format('Y-m-d H:i:s');
-            }
+        foreach ($tokenRevokeEntries as $tokenRevokeEntry) {
             $rows[] = [
-                $tokenResult['t.id'],
-                $tokenResult['u.id'],
-                $tokenResult['userUniqueIdentifier'],
-                $tokenCreated,
-                $tokenExpires ?? '-',
+                $tokenRevokeEntry->getTokenId()->toString(),
+                $tokenRevokeEntry->getUserId()->toString(),
+                $tokenRevokeEntry->getUserUniqueIdentifier(),
+                $tokenRevokeEntry->getTokenCreated()->format('Y-m-d H:i:s'),
+                $tokenRevokeEntry->getTokenExpirationDate()?->format('Y-m-d H:i:s') ?? '-',
             ];
             if (count($rows) >= self::NUMBER_OF_TOKENS_TO_DISPLAY && $showOnlyFirstTokens) {
                 break;
@@ -265,43 +374,26 @@ class TokenRevokeCommand extends Command
         $table->setRows($rows);
         $table->render();
         $this->io->newLine();
+    }
 
-        if (true === $input->getOption(self::OPTION_DRY_RUN)) {
-            $this->io->finalMessage('Dry run finished.');
-
-            return self::SUCCESS;
-        }
-
-        if (true !== $input->getOption(self::OPTION_FORCE)) {
-            /**
-             * @var QuestionHelper $helper
-             */
-            $helper = $this->getHelper('question');
-            $question = new ConfirmationQuestion(sprintf(
-                '  Are you sure you want to revoke all %d tokens? [y/N]: ',
-                $countTokensToBeRevoked
-            ), false);
-            if (!$helper->ask($input, $output, $question)) {
-                $this->io->finalMessage('Aborted revoking tokens.');
-
-                return self::FAILURE;
-            }
-            $this->io->newLine();
-        }
-
-        $this->io->writeln('  Revoking tokens...');
-
-        $progressBar = $this->io->createProgressBarInInteractiveTerminal($countTokensToBeRevoked);
+    /**
+     * @param TokenRevokeEntry[] $tokenRevokeEntries
+     */
+    private function revokeTokens(array $tokenRevokeEntries): void
+    {
+        $countTokenRevokeEntries = count($tokenRevokeEntries);
+        $progressBar = $this->io->createProgressBarInInteractiveTerminal($countTokenRevokeEntries);
         $progressBar?->display();
 
-        foreach ($res as $i => $tokenResult) {
-            $tokenUuid = Uuid::fromString($tokenResult['t.id']);
-            $tokenElement = $this->elementManager->getNode($tokenUuid);
+        $currentItem = 1;
+        foreach ($tokenRevokeEntries as $tokenRevokeEntry) {
+            ++$currentItem;
+            $tokenElement = $this->elementManager->getNode($tokenRevokeEntry->getTokenId());
             if (!$tokenElement) {
                 $message = sprintf(
                     'Unable to revoke token with UUID %s from user with UUID %s, as token could not be fetched from database.',
-                    $tokenResult['t.id'],
-                    $tokenResult['u.id']
+                    $tokenRevokeEntry->getTokenId()->toString(),
+                    $tokenRevokeEntry->getUserId()->toString()
                 );
                 $this->io->writeln($message);
                 $this->logger->notice($message);
@@ -309,18 +401,20 @@ class TokenRevokeCommand extends Command
             }
             $tokenElement->addProperty('state', TokenStateType::REVOKED->value);
             $this->elementManager->merge($tokenElement);
-            if (0 === $i % self::TOKEN_REVOKE_BATCH_SIZE) {
+
+            $this->redisClient->expire(
+                $this->authProvider->getRedisTokenKeyFromHashedToken($tokenRevokeEntry->getTokenHash()),
+                0
+            );
+
+            if (0 === $currentItem % self::TOKEN_REVOKE_BATCH_SIZE) {
                 $this->elementManager->flush();
-                $progressBar?->setProgress($i + 1);
+                $progressBar?->setProgress($currentItem);
             }
         }
         $this->elementManager->flush();
         $progressBar?->finish();
         $progressBar?->clear();
         $this->io->newLine();
-
-        $this->io->finalMessage('Successfully revoked tokens.');
-
-        return Command::SUCCESS;
     }
 }